Extract out generic helper for undefinedDropdownValue
