+++ src/elements/image/__tests__/imageElementDataTransformer.spec.ts
@@ -1,6 +1,6 @@
 import type { FieldNameToValueMap } from "../../../plugin/helpers/fieldView";
+import { undefinedDropdownValue } from "../../helpers/transform";
 import type { createImageFields } from "../ImageElement";
-import { undefinedDropdownValue } from "../ImageElement";
 import type { ImageFields } from "../imageElementDataTransformer";
 import { transformElement } from "../imageElementDataTransformer";
