+++ src/elements/video/videoDataTransformer.ts
@@ -1,4 +1,5 @@
 import type { FieldNameToValueMap } from "../../plugin/helpers/fieldView";
+import { undefinedDropdownValue } from "../helpers/transform";
 import type { TransformIn, TransformOut } from "../helpers/types/Transform";
 import type { createVideoFields } from "./VideoSpec";
 
@@ -13,6 +14,7 @@ export type ExternalVideoFields = {
   title: string;
   html: string;
   authorName: string;
+  role: string | undefined;
 };
 
 export type ExternalVideoData = {
@@ -26,9 +28,10 @@ export type PartialVideoData = {
 export const transformElementIn: TransformIn<
   PartialVideoData,
   ReturnType<typeof createVideoFields>
-> = ({ fields: isMandatory, ...rest }) => {
+> = ({ fields: { isMandatory, role, ...rest } }) => {
   return {
     isMandatory: isMandatory === "true",
+    role: role ?? undefinedDropdownValue,
     ...rest,
   };
 };
@@ -38,6 +41,7 @@ export const transformElementOut: TransformOut<
   ReturnType<typeof createVideoFields>
 > = ({
   isMandatory,
+  role,
   ...rest
 }: FieldNameToValueMap<
   ReturnType<typeof createVideoFields>
@@ -45,6 +49,7 @@ export const transformElementOut: TransformOut<
   return {
     fields: {
       isMandatory: isMandatory.toString(),
+      role: role === undefinedDropdownValue ? undefined : role,
       ...rest,
     },
   };
